% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{.m3_backend}
\alias{.m3_backend}
\title{Which Python backend to use ("basilisk" or a system "reticulate" python).}
\usage{
.m3_backend()
}
\description{
Which Python backend to use ("basilisk" or a system "reticulate" python).
}
\keyword{internal}
