% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{m3_attribution_matrix}
\alias{m3_attribution_matrix}
\title{The full cell x gene signed attribution matrix.}
\usage{
m3_attribution_matrix(attribution)
}
\arguments{
\item{attribution}{an \code{m3_attribution}.}
}
\value{
a numeric matrix, cells x features.
}
\description{
The full cell x gene signed attribution matrix.
}
