% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{m3_capabilities}
\alias{m3_capabilities}
\title{Capabilities of a trained model (embedding / reconstruct / predict_donors).}
\usage{
m3_capabilities(model)
}
\arguments{
\item{model}{an \code{m3_model}.}
}
\value{
a named logical vector.
}
\description{
Capabilities of a trained model (embedding / reconstruct / predict_donors).
}
