% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{m3_dataset_obs}
\alias{m3_dataset_obs}
\title{Per-cell metadata (obs) of a dataset as a data.frame.}
\usage{
m3_dataset_obs(dataset)
}
\arguments{
\item{dataset}{an \code{m3_dataset}.}
}
\value{
a data.frame with one row per cell.
}
\description{
Per-cell metadata (obs) of a dataset as a data.frame.
}
